refactor: display 관련 상수와 메서드를 ui 폴더로 이동
