+++ src/consts.js
@@ -1,7 +1,2 @@
 export const RESULT_SUCCESS = '성공';
 export const RESULT_NOTHING = '낫싱';
-
-export const MSG_SUCCESS = '🎉정답을 맞추셨습니다!🎉';
-export const MSG_RESTART = '게임을 새로 시작하시겠습니까? ';
-
-export const ACTION_RESTART = 'restart';
