+++ src/index.js
@@ -1,10 +1,11 @@
+import displaySuccess from './ui/displaySuccess.js';
+import displayText from './ui/displayText.js';
 import getBallCount from './utils/getBallCount.js';
 import getStrikeCount from './utils/getStrikeCount.js';
 import removeChildren from './utils/removeChildren.js';
 import generateComputerInput from './library/generateComputerInput.js';
-import {
-  RESULT_SUCCESS, RESULT_NOTHING, MSG_SUCCESS, MSG_RESTART, ACTION_RESTART,
-} from './consts.js';
+import { RESULT_SUCCESS, RESULT_NOTHING } from './consts.js';
+import { ACTION_RESTART } from './ui/consts.js';
 
 export default class BaseballGame {
   constructor(app, input, container) {
@@ -30,34 +31,11 @@ export default class BaseballGame {
     return result;
   }
 
-  static displaySuccess(parent) {
-    const heading = document.createElement('h4');
-    heading.innerText = MSG_SUCCESS;
-
-    const typography = document.createElement('span');
-    typography.innerText = MSG_RESTART;
-
-    const button = document.createElement('button');
-    button.textContent = '게임 재시작';
-    button.dataset.action = ACTION_RESTART;
-
-    parent.appendChild(heading);
-    parent.appendChild(typography);
-    parent.appendChild(button);
-  }
-
-  static displayText(parent, text) {
-    const typography = document.createElement('span');
-    typography.innerText = text;
-
-    parent.appendChild(typography);
-  }
-
   displayResult(result) {
     removeChildren(this.container);
 
-    if (result === RESULT_SUCCESS) this.displaySuccess(this.container);
-    else this.displayText(this.container, result);
+    if (result === RESULT_SUCCESS) displaySuccess(this.container);
+    else displayText(this.container, result);
   }
 
   [ACTION_RESTART](e) {
