+++ src/ui/consts.js
@@ -1,0 +1,6 @@
+export const MSG_SUCCESS = '🎉정답을 맞추셨습니다!🎉';
+export const MSG_RESTART = '게임을 새로 시작하시겠습니까? ';
+
+export const ACTION_RESTART = 'restart';
+
+export const RESTART_BUTTON_ID = 'game-restart-button';
